+++ backend/src/neuroagent/tools/entitycore_simulationcampaign_getall.py
@@ -90,6 +90,9 @@ class SimulationCampaignGetAllTool(BaseTool):
         response_data = response.json()
         for simulation in response_data["data"]:
             simulation["assets"] = []
+            simulation["url_link"] = (
+                self.metadata.entity_frontend_url + "/" + simulation["id"]
+            )
         return ListResponseSimulationCampaignRead(**response_data)
 
     @classmethod