+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Possibility to change the OpenAI SDK base url via an env var.
 
 ### Changed
+- Made the context analyzer an url analyzer instead.
+
+### Fixed
+- `url_link` in tools that retrieve explore data.
 - Remove `sonata_asset_id` from circuit analysis tool dependency.
 
 ## [v0.9.2] - 25.09.2025